ree
massivley increased health and market values of technique scrolls
all technique scrolls stack size are now 1 since they dont get destroyed on use anymore.
updated desc of some technique scrolls
removed old bug-testing logs relating to spirit trees

remade qi gather and element emit grid logic. they now recalc all values each time they are changed to keep them consistent, this should fix all current bugs relating to them.

+++ 1.5/Source/Ascension/Ascension_PostHarmony.cs
@@ -23,7 +23,7 @@ namespace Ascension
                     if (compQuality != null)
                     {
                         QualityCategory newQuality = QualityCategory.Awful;
-                        int currentQi = cauldronComp.currentQi;
+                        float currentQi = cauldronComp.currentQi;
 
                         // Determine the highest quality allowed by current Qi
                         for (int i = QualityQiThresholds.Length - 1; i > 0; i--)
